fix: update editUrl to working link
Not sure if this will fix the 404's, but worth a shot. I'm thinking the fact the issue only exists on subpages and not the index is because the index is loaded directly from `src/components`, while docusaurus has a bigger impact elsewhere. So the issue may reside in the config file.