Update deployment-plans.mdx
Fixes a 404 error to the "test contract" link in the additional resources.

Updates language in additional resources for better Clarity

Fixes broken anchor link

Note: @ryanwaits  in a callout near the bottom there is a reference to a "Check contract" guide that just links to this same page. I'm not sure what that should point to, but I'm fine with simply removing that sentence altogether.